+++ src/com/github/mjkuranda/retrographicsmanager/commandsystem/commands/ConvertCommand.java
@@ -1,6 +1,12 @@
 package com.github.mjkuranda.retrographicsmanager.commandsystem.commands;
 
+import com.github.mjkuranda.retrographicsmanager.converters.ImageConverter;
+
+import java.awt.image.BufferedImage;
 import java.io.File;
+import java.io.IOException;
+import java.nio.file.Files;
+import java.nio.file.Paths;
 
 public class ConvertCommand implements Command {
 
@@ -12,9 +18,15 @@ public class ConvertCommand implements Command {
 
     @Override
     public void execute() {
-        File originalImage = new File(lineArgs[2]);
-        File convertedImage = new File(lineArgs[3]);
+        if (!new File("dat/output/conversions").exists()) {
+            try {
+                Files.createDirectories(Paths.get("dat/output/conversions"));
+            } catch (IOException e) {
+                e.printStackTrace();
+            }
+        }
 
-
+        BufferedImage img = ImageConverter.toRetroImage(lineArgs[2], "light".equals(lineArgs[1]));
+        ImageConverter.saveImage(img, lineArgs[3]);
     }
 }